requirements.txt
python-telegram-bot>=20.7
python-binance>=1.0.19
pandas>=2.1.0
numpy>=1.24.0
ta>=0.10.0
ccxt>=4.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
websocket-client>=1.6.0
